requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.4
exceptiongroup==1.0.4
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-xdist==3.0.2
requests==2.27.1
selenium==4.6.1
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
allure-pytest==2.12.0
allure-python-commons==2.12.0
pyotp==2.8.0
